Some code improvements.

+++ src/main/java/ar/com/tacsutn/grupo1/eventapp/repositories/EventListRepository.java
@@ -11,5 +11,5 @@ public interface EventListRepository extends CrudRepository<EventList, Long> {
 
     Page<EventList> findAllByUser(User user, Pageable pageable);
 
-    Integer countAllByUserId(long user_id);
+    long countAllByUserId(long user_id);
 }
